Better accomodate builds triggered by PRs
PR builds are different from builds on a branch because they build
the result of a hypothetical merge with master. Committing the result of that
build to the source branch's analog in the website-builds repo is a little
weird, but seems least likely to make a mess in the website-builds repo.

This commit makes the following changes:
- Commit PR builds to the source branch instead of master
- Leave a helpful commit message referencing the original PR
- Set HEAD to the source branch manually rather than relying on `git checkout`