add note about package aliases
This satisfies the concerns in #263 without sacrificing forward-compatibility and
keeping things shorter to type and easier to remember.

The partial is added to Fedora and CentOS/RHEL where it currently applies, but
could be added to other distros where it is relevant.

The package name prefix used by each distro is defined in
{{python_prefix}}, and the note can be switched off by making that falsy. This
is done when just certbot is installed, since that isn't an alias.