Set hyphenated url in js using history
When a js user visits an instruction page using the dropdowns, set the url using the history api. When they hit back, use the history.onpopstate to update the page appropriately.

We no longer need to parse the url to update the page outside of forward/back actions because we will jekyll will render those pages individually.