Fix Namecheap hosting tutorial
Arguably, Namecheap should be categorized as "no": their support for
Let's Encrypt is materially identical to that of GoDaddy's -- currently
categorized as "no" -- they provide access to cPanel's SSL manager, and
have no support for ACME in any official capacity. It's not even clear
whether the certificates CAN be *updated* in an automated way if the
provider's paid product isn't used (I am currently testing this myself.)

The page previously-linked "link" was NOT a link to the webhosting pro-
vider's main page; rather, it was a link to Namecheap's partner/in-house
TLS CA cert service, which directly "competes" with Let's Encrypt while
offering no free plans.

I have replaced that link with their main hosting-services page.

The page previously-linked "tutorial" was NOT instructions on how to
configure Let's Encrypt with their hosting; it was (crucially) _not even
applicable_ to "third-party product[s]", such as Let's Encrypt certs. It
was, rather, a promotional/instructional page that explained ONLY how to
configure the free trial of their partner CA's certificates.

I have replaced that link with what little official documentation they
provide on how to install+configure (via cPanel) "third-party" TLS
certificates (such as Let's Encrypt's).

The page previously-linked "plan" was NOT any kind of plans or press
release from Namecheap regarding Let's Encrypt support; rather, it was
simply the same "link" earlier to their paid CA service.

I have replaced that link with the latest statement from Namecheap
regarding Let's Encrypt support. It's quite bleak, and had to be relayed
by a third-party due to lack of any _officially_-published press
releases or documentation; nevertheless, it's the best we have, and the
previous content of that field was _strictly_ irrelevant.