Move instructions for Arch Linux, openSUSE, and other UNIX to snaps (#648)
This fixes certbot/certbot#8291.

As described in that issue and the Google Doc linked there, I changed the "Other UNIX" instructions to "Other Linux". I kept the id as pip to avoid breaking links to those instructions. I manually tested this and all of the dropdown options we expect display the snap instructions.

I initially was blocking doing this on having (part of) certbot/certbot#8280 resolved, but a new enough version of snapd is packaged for both Arch Linux and openSUSE and I think getting people testing these instructions is important since the current plan is to deprecate certbot-auto on these platforms in our next release. I wrote more about the timeline here at certbot/certbot#8292 if you're interested.

* Move generic, arch linux, and suse to snaps.

* Rename "Other UNIX" to "Other Linux"