macos: use $(command -v certbot) for cron (#729)
Certbot is not in the PATH when installed from Homebrew.

---

Per https://github.com/certbot/certbot/issues/8974.

Excuse the templating. I couldn't work out a way to do this in the "controller" without refactoring a bunch of other stuff. Any bright ideas on not having this be so hacky would be appreciated.

After this change, the macOS command becomes:

    echo "0 0,12 * * * root python -c 'import random; import time; time.sleep(random.random() * 3600)' && $(command -v certbot) renew -q" | sudo tee -a /etc/crontab > /dev/null

and the Linux (e.g. `pip`) command remains as:

    echo "0 0,12 * * * root /opt/certbot/bin/python -c 'import random; import time; time.sleep(random.random() * 3600)' && certbot renew -q" | sudo tee -a /etc/crontab > /dev/null