Update hosting_providers.json (#775)
* Update hosting_providers.json

Our customers are big fan of Let's Encrypt and as a hosting provider our wish is also to be the part of the shout outs on eff website. I have added our Snel.com as a hosting company to your website. In awaiting for your approval :)

* Update hosting_providers.json

Links will be update

* Update hosting_providers.json

* Update hosting_providers.json

* Update hosting_providers.json