Update hosting_providers.json
Just above the subtitle "¿Qué es Hosting?" The first logo of Let's Encrypt is shown.  The tutorisl refers to rhe AutoSSL button of Let's Encrypt module of cOanel, that actívates and renews periodicslly all certs with possibility to exclude selected domains from having SSL.

Subtle info of Let's Encrypt is givenbeyond the logo,   because sale of paid Certs is explicitly promoted.  

But open support is given to use free let's encrypt for everyone. The module AutoSSL of cPanel has type of cert option as "Let's encrypt" as default shown in image in this branch, which user has access when logged onto cPanel